requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.41 
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Authlib==1.2.1
Werkzeug==2.0.3
python-dotenv==0.20.0
numpy==1.21.6
pandas==1.3.5
mysql-connector-python==8.0.28
requests
pymysql